Revert doing input through pty; too strange.
Delete CancellableQueueIterator; it didn't make sense to have cancellation
be based on high-frequency polling if the queue was empty.  On cancel, simply
put None in the queue.

Use special handling for input to first command, for now.

Always use /bin/sh instead of passing shell=True; we can't reliably use
the user's preferred shell as quoting etc. may be different.




git-svn-id: https://hotwire-shell.googlecode.com/svn/trunk@227 79f8d753-db37-0410-980a-393e3d9e3e3e